refactor: 스크롤 복원 유틸 공통화 #1656
